Moved the carcass module stuff into a script w/in the animal module as I totally misunderstood the review.
