+++ tests/test_carcasses.py
@@ -1,6 +1,6 @@
-"""Test module for carcass_module.py.
+"""Test module for carcasses.py.
 
-This module tests the functionality of dummy_animal_module.py
+This module tests the functionality of carcasses.py
 """
 
 
@@ -9,7 +9,7 @@ class TestCarcassPool:
 
     def test_initialization(self):
         """Testing initialization of CarcassPool."""
-        from virtual_rainforest.models.carcasses.carcass_module import CarcassPool
+        from virtual_rainforest.models.animals.carcasses import CarcassPool
 
         c1 = CarcassPool(1000.7, 1)
         assert c1.energy == 1000.7
