+++ virtual_rainforest/models/animals/carcasses.py
